requirements.txt
# HTR Multi-Scale Contrastive Learning - Requirements
# Optimized for NVIDIA A40 and Python 3.8+

# Core PyTorch (install separately for CUDA support)
# torch>=2.0.0
# torchvision>=0.15.0
# torchaudio>=2.0.0

# Numerical computing
numpy>=1.21.0
scipy>=1.7.0

# Computer vision and image processing
opencv-python>=4.5.0
scikit-image>=0.19.0
Pillow>=8.3.0

# Machine learning utilities
scikit-learn>=1.0.0

# Data visualization and plotting
matplotlib>=3.5.0
seaborn>=0.11.0

# Data handling
pandas>=1.3.0

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.60.0
pathlib2>=2.3.5

# Logging and monitoring
tensorboard>=2.8.0

# Optional: Advanced experiment tracking
# wandb>=0.12.0

# Development and testing (optional)
# pytest>=6.0.0
# black>=22.0.0
# isort>=5.0.0
# flake8>=4.0.0

# Jupyter notebook support (optional)
# jupyter>=1.0.0
# ipywidgets>=7.6.0